# docs/apifox/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "网上商城系统 API",
    "description": "Spring Boot 2.7 + MyBatis-Plus + MySQL + Redis + JWT。管理端前缀 /admin，用户端前缀 /user；除登录接口外均需在 Header 携带 token。本文件由 tools/export_openapi.py 从 Controller 源码自动提取生成，与实现同源。",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "本地环境"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "token": {
        "type": "apiKey",
        "in": "header",
        "name": "token",
        "description": "登录接口返回的 JWT，放在请求头 token 字段"
      }
    },
    "schemas": {
      "LoginDTO": {
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "username": {
            "type": "string",
            "example": "zhangsan"
          },
          "password": {
            "type": "string",
            "example": "123456"
          }
        }
      },
...
      "DishDTO": {
        "type": "object",
        "required": [
          "name",
          "categoryId",
          "price"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string",
            "example": "机械键盘"
          },
          "categoryId": {
            "type": "integer",
            "example": 2
          },
          "price": {
            "type": "number",
            "example": 32.0
          },
          "image": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "description": "0停售 1起售",
            "example": 1
          },
          "flavors": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
...
      "EmployeeDTO": {
        "type": "object",
        "required": [
          "name",
          "username",
          "password"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "张小厨"
          },
          "username": {
            "type": "string",
            "example": "kitchen01"
          },
          "password": {
            "type": "string",
            "example": "Test@12345"
          },
          "phone": {
            "type": "string",
            "description": "选填，^1[3-9]\\d{9}$",
            "example": "13900001111"
          }
        }
      },
      "EmployeeUpdateDTO": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "张小厨"
          },
          "phone": {
            "type": "string",
            "description": "选填，^1[3-9]\\d{9}$",
            "example": "13900001111"
          }
        }
      }
    }
  },
  "security": [
    {
      "token": []
    }
  ],
  "paths": {
...
    "/admin/dish": {
      "post": {
        "tags": [
          "商品管理"
        ],
        "summary": "save",
        "operationId": "DishController_save",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DishDTO"
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "商品管理"
        ],
        "summary": "update",
        "operationId": "DishController_update",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DishDTO"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "商品管理"
        ],
        "summary": "deleteBatch",
        "operationId": "DishController_deleteBatch",
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/admin/dish/list": {
      "get": {
        "tags": [
          "商品管理"
        ],
        "summary": "listByCategoryId",
        "operationId": "DishController_listByCategoryId",
        "parameters": [
          {
            "name": "categoryId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/admin/dish/page": {
      "get": {
        "tags": [
          "商品管理"
        ],
        "summary": "page",
        "operationId": "DishController_page",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/admin/dish/status/{status}": {
      "post": {
        "tags": [
          "商品管理"
        ],
        "summary": "startOrStop",
        "operationId": "DishController_startOrStop",
        "parameters": [
          {
            "name": "status",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/admin/employee": {
      "post": {
        "tags": [
          "员工管理"
        ],
        "summary": "save",
        "operationId": "EmployeeController_save",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmployeeDTO"
              }
            }
          }
        }
      }
    },
    "/admin/employee/login": {
      "post": {
        "tags": [
          "员工管理"
        ],
        "summary": "login",
        "operationId": "EmployeeController_login",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDTO"
              }
            }
          }
        },
        "security": []
      }
    },
    "/admin/employee/page": {
      "get": {
        "tags": [
          "员工管理"
        ],
        "summary": "page",
        "operationId": "EmployeeController_page",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/admin/employee/{id}": {
      "put": {
        "tags": [
          "员工管理"
        ],
        "summary": "update",
        "operationId": "EmployeeController_update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmployeeUpdateDTO"
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "员工管理"
        ],
        "summary": "getById",
        "operationId": "EmployeeController_getById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/admin/employee/{id}/status/{status}": {
      "put": {
        "tags": [
          "员工管理"
        ],
        "summary": "startOrStop",
        "operationId": "EmployeeController_startOrStop",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "status",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/admin/order/accept/{id}": {
      "put": {
        "tags": [
          "订单管理"
        ],
        "summary": "accept",
        "operationId": "OrderController_accept",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/admin/order/complete/{id}": {
      "put": {
        "tags": [
          "订单管理"
        ],
        "summary": "complete",
        "operationId": "OrderController_complete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/admin/order/details/{id}": {
      "get": {
        "tags": [
          "订单管理"
        ],
        "summary": "details",
        "operationId": "OrderController_details",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/admin/order/page": {
      "get": {
        "tags": [
          "订单管理"
        ],
        "summary": "page",
        "operationId": "OrderController_page",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "number",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "beginTime",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endTime",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/admin/order/reject/{id}": {
      "put": {
        "tags": [
          "订单管理"
        ],
        "summary": "reject",
        "operationId": "OrderController_reject",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "reason",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/user/dish/list": {
      "get": {
        "tags": [
          "商品浏览"
        ],
        "summary": "list",
        "operationId": "UserDishController_list",
        "parameters": [
          {
            "name": "categoryId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  }
}